ImageParser/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anytree==2.8.0
appdirs==1.4.3
astor==0.8.1
attrs==19.3.0
autopep8==1.4.4
awscli==1.16.310
Babel==2.8.0
Beaker==1.11.0
beautifulsoup4==4.8.1
botocore==1.13.46
bs4==0.0.1
btrfsutil==1.1.1
CacheControl==0.12.6
cachetools==3.1.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
contextlib2==0.6.0
cryptocompare==0.6.4
cupshelpers==1.0
cycler==0.10.0
dill==0.3.1.1
distlib==0.3.0
distro==1.4.0
docopt==0.6.2
docutils==0.14
fix-yahoo-finance==0.1.37
Flask==1.1.1
future==0.18.2
gast==0.2.2
google-auth==1.8.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.25.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lensfun==0.3.2
lit==0.9.1.dev0
lockfile==0.12.2
louis==3.12.0
lxml==4.4.2
Mako==1.1.0
mallard-ducktype==1.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
msgpack==0.6.2
multitasking==0.0.9
netifaces==0.10.9
nomics-python==2.0.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
ordered-set==3.1.1
packaging==20.0
pandas==0.25.3
pandas-datareader==0.8.1
pep517==0.6.0
Pillow==7.0.0
pkgconfig==1.5.1
progress==1.5
progressbar2==3.47.0
promise==2.3
protobuf==3.11.2
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycairo==1.18.2
pycodestyle==2.5.0
pycups==1.9.74
pycurl==7.43.0.3
Pygments==2.5.2
PyGObject==3.34.0
pylast==3.2.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.0
python-dateutil==2.8.1
python-mnist==0.6
python-utils==2.3.0
pytoml==0.1.21
pytz==2019.3
pyxdg==0.26
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rpm==4.15.1
rsa==4.0
s-tui==0.8.3
s3transfer==0.2.1
scipy==1.4.1
selenium==3.141.0
setuptools-scm==3.3.3
sip==4.19.20
six==1.13.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
sportmonks==1.1.0
SQLAlchemy==1.3.12
team==1.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-datasets==2.0.0
tensorflow-estimator==2.1.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
tqdm==4.42.1
tzlocal==2.0.0
urllib3==1.25.7
urwid==2.1.0
virtualenv==16.1.0
virtualfish==1.0.6
webencodings==0.5.1
Werkzeug==0.16.0
Whoosh==2.7.4
wrapt==1.11.2
yfinance==0.1.52
youtube-dl==2020.1.15
zipp==0.6.0
